+++ SleepModel/HEMSleepSoundConfigurationCell.m
@@ -7,15 +7,39 @@
 //
 
 #import "HEMSleepSoundConfigurationCell.h"
+#import "HEMStyle.h"
+
+static CGFloat const HEMSleepSoundConfCellSeparatorHeight = 0.5f;
+
+@interface HEMSleepSoundConfigurationCell()
+
+@property (nonatomic, weak) UIView* overlay;
+
+@end
 
 @implementation HEMSleepSoundConfigurationCell
 
-/*
-// Only override drawRect: if you perform custom drawing.
-// An empty implementation adversely affects performance during animation.
-- (void)drawRect:(CGRect)rect {
-    // Drawing code
+- (void)awakeFromNib {
+    [super awakeFromNib];
+    [[self titleSeparator] setBackgroundColor:[UIColor separatorColor]];
+    [[self soundSeparator] setBackgroundColor:[UIColor separatorColor]];
+    [[self durationSeparator] setBackgroundColor:[UIColor separatorColor]];
+    
+    [[self titleSeparatorHeight] setConstant:HEMSleepSoundConfCellSeparatorHeight];
+    [[self soundSeparatorHeight] setConstant:HEMSleepSoundConfCellSeparatorHeight];
+    [[self durationSeparatorHeight] setConstant:HEMSleepSoundConfCellSeparatorHeight];
 }
-*/
+
+- (void)deactivate:(BOOL)deactivate {
+    [self setUserInteractionEnabled:!deactivate];
+    if (deactivate) {
+        UIView* overlay = [[UIView alloc] initWithFrame:[self bounds]];
+        [overlay setBackgroundColor:[UIColor colorWithWhite:1.0f alpha:0.5f]];
+        [[self contentView] addSubview:overlay];
+        [self setOverlay:overlay];
+    } else {
+        [[self overlay] removeFromSuperview];
+    }
+}
 
 @end