+++ SleepModel/HEMSleepSoundDurationsPresenter.h
@@ -1,0 +1,13 @@
+//
+//  HEMSleepSoundDurationsPresenter.h
+//  Sense
+//
+//  Created by Jimmy Lu on 3/25/16.
+//  Copyright © 2016 Hello. All rights reserved.
+//
+
+#import "HEMListPresenter.h"
+
+@interface HEMSleepSoundDurationsPresenter : HEMListPresenter
+
+@end