+++ SleepModel/HEMSleepSoundPlayerPresenter.m
@@ -13,6 +13,7 @@
 #import "HEMSleepSoundConfigurationCell.h"
 #import "HEMSleepSoundService.h"
 #import "HEMActivityIndicatorView.h"
+#import "HEMStyle.h"
 
 static CGFloat const HEMSleepSoundConfigCellHeight = 217.0f;
 
@@ -27,22 +28,28 @@ typedef NS_ENUM(NSInteger, HEMSleepSoundPlayerState) {
 @property (nonatomic, weak) HEMSleepSoundService* service;
 @property (nonatomic, weak) UICollectionView* collectionView;
 @property (nonatomic, weak) UIButton* actionButton;
+@property (nonatomic, strong) SENSleepSounds* cachedSounds;
 @property (nonatomic, strong) SENSleepSounds* availableSounds;
 @property (nonatomic, strong) SENSleepSoundDurations* availableDurations;
 @property (nonatomic, assign, getter=isLoading) BOOL loading;
 @property (nonatomic, strong) HEMActivityIndicatorView* indicatorView;
+@property (nonatomic, strong) SENSleepSound* selectedSound;
+@property (nonatomic, strong) SENSleepSoundDuration* selectedDuration;
 
 // TODO: remove once we hook everything up.  We should check status upon load
 @property (nonatomic, assign) HEMSleepSoundPlayerState playerState;
+@property (nonatomic, weak) HEMSleepSoundConfigurationCell* configCell;
 
 @end
 
 @implementation HEMSleepSoundPlayerPresenter
 
-- (instancetype)initWithSleepSoundService:(HEMSleepSoundService*)service {
+- (instancetype)initWithSleepSoundService:(HEMSleepSoundService*)service
+                           andSleepSounds:(nullable SENSleepSounds*)sleepSounds {
     self = [super init];
     if (self) {
         _service = service;
+        _cachedSounds = sleepSounds;
     }
     return self;
 }
@@ -53,6 +60,7 @@ typedef NS_ENUM(NSInteger, HEMSleepSoundPlayerState) {
     itemSize.height = HEMSleepSoundConfigCellHeight;
     [layout setItemSize:itemSize];
     
+    [collectionView setAlwaysBounceVertical:YES];
     [collectionView setDataSource:self];
     [collectionView setDelegate:self];
     [self setCollectionView:collectionView];
@@ -99,10 +107,16 @@ typedef NS_ENUM(NSInteger, HEMSleepSoundPlayerState) {
     
     dispatch_group_t dataGroup = dispatch_group_create();
     
-    dispatch_group_enter(dataGroup);
-    [self loadSleepSounds:^{
-        dispatch_group_leave(dataGroup);
-    }];
+    // might have been provided to the presenter already
+    if (![self cachedSounds]) {
+        dispatch_group_enter(dataGroup);
+        [self loadSleepSounds:^{
+            dispatch_group_leave(dataGroup);
+        }];
+    } else {
+        [self setAvailableSounds:[self cachedSounds]];
+        [self setCachedSounds:nil]; // remove the cache
+    }
     
     dispatch_group_enter(dataGroup);
     [self loadSleepSoundDurations:^{
@@ -111,13 +125,22 @@ typedef NS_ENUM(NSInteger, HEMSleepSoundPlayerState) {
     
     __weak typeof(self) weakSelf = self;
     dispatch_group_notify(dataGroup, dispatch_get_main_queue(), ^{
-        [[weakSelf collectionView] reloadData];
-        [weakSelf setLoading:NO];
+        __strong typeof(weakSelf) strongSelf = weakSelf;
+        HEMSleepSoundService* service = [strongSelf service];
+        if (![strongSelf selectedSound]) {
+            [strongSelf setSelectedSound:[service defaultSleepSoundFrom:[strongSelf availableSounds]]];
+        }
+        if (![strongSelf selectedDuration]) {
+            [strongSelf setSelectedDuration:[service defaultDurationFrom:[strongSelf availableDurations]]];
+        }
+        [[strongSelf collectionView] reloadData];
+        [strongSelf setLoading:NO];
     });
 }
 
 - (void)setPlayerState:(HEMSleepSoundPlayerState)playerState {
     _playerState = playerState;
+    [[self configCell] deactivate:NO];
     [[self actionButton] setEnabled:YES];
     [[self indicatorView] stop];
     [[self indicatorView] removeFromSuperview];
@@ -132,6 +155,7 @@ typedef NS_ENUM(NSInteger, HEMSleepSoundPlayerState) {
                                  forState:UIControlStateNormal];
             break;
         default: {
+            [[self configCell] deactivate:YES];
             [[self actionButton] setEnabled:NO];
             
             if (![self indicatorView]) {
@@ -151,6 +175,26 @@ typedef NS_ENUM(NSInteger, HEMSleepSoundPlayerState) {
             [[self actionButton] setImage:nil forState:UIControlStateNormal];
             [[self actionButton] addSubview:[self indicatorView]];
             break;
+        }
+    }
+}
+
+- (void)setSelectedSound:(SENSleepSound*)sound {
+    BOOL shouldReload = _selectedSound != nil;
+    if (![[_selectedSound identifier] isEqualToNumber:[sound identifier]]) {
+        _selectedSound = sound;
+        if (shouldReload) {
+            [[self collectionView] reloadData];
+        }
+    }
+}
+
+- (void)setSelectedDuration:(SENSleepSoundDuration*)duration {
+    BOOL shouldReload = _selectedDuration != nil;
+    if (![[_selectedDuration identifier] isEqualToNumber:[duration identifier]]) {
+        _selectedDuration = duration;
+        if (shouldReload) {
+            [[self collectionView] reloadData];
         }
     }
 }
@@ -179,11 +223,28 @@ typedef NS_ENUM(NSInteger, HEMSleepSoundPlayerState) {
 #pragma mark - UICollectionViewDelegate
 
 - (void)configureSleepSoundConfigurationCell:(HEMSleepSoundConfigurationCell*)cell {
-    SENSleepSound* defaultSound = [[self service] defaultSleepSoundFrom:[self availableSounds]];
-    SENSleepSoundDuration* defaultDuration = [[self service] defaultDurationFrom:[self availableDurations]];
-    [[cell soundValueLabel] setText:[defaultSound localizedName]];
-    [[cell durationValueLabel] setText:[defaultDuration localizedName]];
+    [[cell titleLabel] setTextColor:[UIColor sleepSoundPlayerTitleColor]];
+    [[cell soundLabel] setTextColor:[UIColor sleepSoundPlayerTitleColor]];
+    [[cell soundValueLabel] setText:[[self selectedSound] localizedName]];
+    [[cell soundValueLabel] setTextColor:[UIColor sleepSoundPlayerOptionValueColor]];
+    [[cell soundSelectorButton] addTarget:self
+                                   action:@selector(changeSound:)
+                         forControlEvents:UIControlEventTouchUpInside];
+    [[cell durationLabel] setTextColor:[UIColor sleepSoundPlayerTitleColor]];
+    [[cell durationValueLabel] setText:[[self selectedDuration] localizedName]];
+    [[cell durationValueLabel] setTextColor:[UIColor sleepSoundPlayerOptionValueColor]];
+    [[cell durationSelectorButton] addTarget:self
+                                      action:@selector(changeDuration:)
+                            forControlEvents:UIControlEventTouchUpInside];
+    [[cell volumeLabel] setTextColor:[UIColor sleepSoundPlayerTitleColor]];
     [[cell volumeValueLabel] setText:NSLocalizedString(@"sleep-sounds.volume.high", nil)];
+    [[cell volumeValueLabel] setTextColor:[UIColor sleepSoundPlayerOptionValueColor]];
+    [[cell volumeSelectorButton] addTarget:self
+                                    action:@selector(changeVolume:)
+                          forControlEvents:UIControlEventTouchUpInside];
+    [[cell volumeAccessoryView] setHidden:YES];
+    
+    [self setConfigCell:cell];
 }
 
 - (void)collectionView:(UICollectionView *)collectionView
@@ -195,6 +256,28 @@ typedef NS_ENUM(NSInteger, HEMSleepSoundPlayerState) {
 }
 
 #pragma mark - Player Actions
+
+- (void)changeSound:(UIButton*)button {
+    DDLogVerbose(@"change sound");
+    [[self delegate] showAvailableSounds:[[self availableSounds] sounds]
+                       selectedSoundName:[[self selectedSound] localizedName]
+                               withTitle:NSLocalizedString(@"sleep-sounds.option.title.sound", nil)
+                                subTitle:NSLocalizedString(@"sleep-sounds.option.subtitle.sound", nil)
+                                    from:self];
+}
+
+- (void)changeDuration:(UIButton*)button {
+    DDLogVerbose(@"change duration");
+    [[self delegate] showAvailableDurations:[[self availableDurations] durations]
+                       selectedDurationName:[[self selectedDuration] localizedName]
+                                  withTitle:NSLocalizedString(@"sleep-sounds.option.title.duration", nil)
+                                   subTitle:NSLocalizedString(@"sleep-sounds.option.subtitle.duration", nil)
+                                       from:self];
+}
+
+- (void)changeVolume:(UIButton*)button {
+    DDLogVerbose(@"change volume");
+}
 
 - (void)takeAction:(UIButton*)button {
     switch ([self playerState]) {
@@ -213,22 +296,23 @@ typedef NS_ENUM(NSInteger, HEMSleepSoundPlayerState) {
     DDLogVerbose(@"attempting to play sound");
     [self setPlayerState:HEMSleepSoundPlayerStateWaiting];
     
-    SENSleepSound* sound = [[self service] defaultSleepSoundFrom:[self availableSounds]];
-    SENSleepSoundDuration* duration = [[self service] defaultDurationFrom:[self availableDurations]];
-    
     __weak typeof(self) weakSelf = self;
-    [[self service] playSound:sound forDuration:duration withVolume:100 completion:^(NSError * _Nullable error) {
-        __strong typeof(weakSelf) strongSelf = weakSelf;
-        
-        if (!error) {
-            DDLogVerbose(@"playing sound");
-            [strongSelf setPlayerState:HEMSleepSoundPlayerStatePlaying];
-        } else {
-            DDLogVerbose(@"failed to play sound");
-            [strongSelf setPlayerState:HEMSleepSoundPlayerStateStopped];
-            [[strongSelf delegate] presentError:error];
-        }
-    }];
+    [[self service] playSound:[self selectedSound]
+                  forDuration:[self selectedDuration]
+                   withVolume:80
+                   completion:^(NSError * _Nullable error) {
+                       __strong typeof(weakSelf) strongSelf = weakSelf;
+                    
+                       if (!error) {
+                           DDLogVerbose(@"playing sound");
+                           [strongSelf setPlayerState:HEMSleepSoundPlayerStatePlaying];
+                       } else {
+                           DDLogVerbose(@"failed to play sound");
+                           [strongSelf setPlayerState:HEMSleepSoundPlayerStateStopped];
+                           [[strongSelf delegate] presentError:error];
+                       }
+                       
+                   }];
 }
 
 - (void)stop:(UIButton*)button {