+++ SleepModel/HEMSleepSoundViewController.h
@@ -9,8 +9,11 @@
 #import <UIKit/UIKit.h>
 #import "HEMBaseController.h"
 
+@class SENSleepSounds;
+
 @interface HEMSleepSoundViewController : HEMBaseController
 
 @property (nonatomic, assign, getter=isCancellable) BOOL cancellable;
+@property (nonatomic, strong) SENSleepSounds* sleepSounds;
 
 @end