+++ SleepModel/HEMSleepSoundsPresenter.h
@@ -1,0 +1,13 @@
+//
+//  HEMSleepSoundsPresenter.h
+//  Sense
+//
+//  Created by Jimmy Lu on 3/25/16.
+//  Copyright © 2016 Hello. All rights reserved.
+//
+
+#import "HEMListPresenter.h"
+
+@interface HEMSleepSoundsPresenter : HEMListPresenter
+
+@end